% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{avg_exp_for_complexes}
\alias{avg_exp_for_complexes}
\title{Average Complex Expression}
\usage{
avg_exp_for_complexes(exp_mat, complexes_list)
}
\arguments{
\item{exp_mat}{expression matrix of ligands that may or may not be in a complex}

\item{complexes_list}{named list. names = ligands, elements = ligands and/or complex compononents}
}
\value{
list
}
\description{
Average complex expression
}
\details{
Search each element of complexes_list and determine if actually part of a complex.
If it is, calculate colMeans for all ligands in the complex. If not, return the ligand expression.
rewrite with purrr?
}
