% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{avg_reclig_expr}
\alias{avg_reclig_expr}
\title{Average Receptor/Ligand Expression}
\usage{
avg_reclig_expr(dom, cluster = NULL, genes, complexes)
}
\arguments{
\item{dom}{domino object}

\item{cluster}{vector of clusters to include (NULL uses all)}

\item{genes}{genes to get averages of}

\item{complexes}{list of resolved complexes}
}
\value{
foo
}
\description{
Create a matrix of average expression of cells in provided clsuters
}
\details{
foo
}
