% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{build_domino}
\alias{build_domino}
\title{Calculate a signaling network for a domino object}
\usage{
build_domino(
  dom,
  max_tf_per_clust = 5,
  min_tf_pval = 0.01,
  max_rec_per_tf = 5,
  rec_tf_cor_threshold = 0.15
)
}
\arguments{
\item{dom}{Domino object from create_domino.}

\item{max_rec_per_tf}{Maximum number of receptors to link to each transcription factor.}

\item{rec_tf_cor_threshold}{Minimum pearson correlation used to consider a receptor linked with a transcription factor. Increasing this will decrease the number of receptors linked to each transcription factor.}
}
\value{
A domino object
}
\description{
This function calculates a signaling network. It requires a domino object
preprocessed from create_domino and returns a domino object prepared for 
plotting with the various plotting functions in this package.
}
