% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache}
\alias{cache}
\title{Specify the cache}
\usage{
cache(path = tools::R_user_dir(package = "domino2"))
}
\arguments{
\item{path}{Path to tstore cache on}
}
\value{
BiocFileCache
}
\description{
Specify the cache
}
