% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{circos_receptor_ligand}
\alias{circos_receptor_ligand}
\title{Plot expression of a ligand's receptors by other cell types as a chord plot}
\usage{
circos_receptor_ligand(
  dom,
  ligand,
  receptor_expression_threshold = 0.01,
  cell_idents = NULL,
  cell_colors = NULL,
  title = paste0(ligand, " Signaling"),
  label_arcs = T
)
}
\arguments{
\item{dom}{Domino object that has undergone network building with \code{\link[=build_domino]{build_domino()}}}

\item{ligand}{Name of a ligand active in at least one cell type in the domino object}

\item{receptor_expression_threshold}{Minimum mean expression value of a receptor by a cell type for a chord to be rendered between the cell type and the ligand}

\item{cell_idents}{Vector of cell types from cluster assignments in the domino object to be included in the plot.}

\item{cell_colors}{Named vector of color names or hex codes where names correspond to the plotted cell types and the color values}
}
\value{
renders a circos plot to the active graphics device
}
\description{
Creates a chord plot of expression of receptors that can be activated by a specified
ligand where chord widths correspond to mean receptor expression by the cluster.
}
\examples{
#basic usage
circos_ligand_receptor(domino2:::pbmc_dom_built_tiny, ligand = "FASLG")
#specify colors
cols = c("red", "orange", "green", "blue", "pink", "purple", "slategrey", "firebrick", "hotpink")
names(cols) = levels(domino2:::pbmc_dom_built_tiny@clusters)
circos_receptor_ligand(domino2:::pbmc_dom_built_tiny, ligand = "FASLG", cell_colors = cols)

}
