% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conv_py_bools}
\alias{conv_py_bools}
\title{Change cases of True/False syntax from Python to TRUE/FALSE R syntax}
\usage{
conv_py_bools(obj)
}
\arguments{
\item{obj}{Object that will be converted}
}
\value{
obj The converted object
}
\description{
Change cases of True/False syntax from Python to TRUE/FALSE R syntax
}
\keyword{internal}
