% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Create a heatmap of features organized by cluster}
\usage{
cor_heatmap(
  dom,
  bool = TRUE,
  bool_thresh = 0.15,
  title = TRUE,
  feats = NULL,
  recs = NULL,
  mark_connections = FALSE,
  ...
)
}
\arguments{
\item{dom}{A domino object with network built (build_domino)}

\item{bool}{A boolean indicating whether the heatmap should be continuous or boolean. If boolean then bool_thresh will be used to determine how to define activity as positive or negative.}

\item{bool_thresh}{A numeric indicating the threshold separating 'on' or 'off' for feature activity if making a boolean heatmap.}

\item{title}{Either a string to use as the title or a boolean describing whether to include a title. In order to pass the 'main' parameter to NMF::aheatmap you must set title to FALSE.}

\item{feats}{Either a vector of features to include in the heatmap or 'all' for all features. If left NULL then the features selected for the signaling network will be shown.}

\item{recs}{Either a vector of receptors to include in the heatmap or 'all' for all receptors. If left NULL then the receptors selected in the signaling network connected to the features plotted will be shown.}

\item{mark_connections}{A boolean indicating whether to add an 'x' in cells where there is a connected receptor or TF. Default FALSE.}

\item{...}{Other parameters to pass to NMF::aheatmap. Note that to use the 'main' parameter of NMF::aheatmap you must set title = FALSE and to use 'annCol' or 'annColors' ann_cols must be FALSE.}
}
\description{
Creates a heatmap of feature expression (typically transcription factor
activation scores) by cells organized by cluster.
}
