% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{correlation_grid}
\alias{correlation_grid}
\title{Create correlation plots for features and receptors}
\usage{
correlation_grid(dom, recs, feats)
}
\arguments{
\item{dom}{A domino object with network built (build_domino)}

\item{recs}{A vector of receptors to create correlation plots for}

\item{feats}{A vector of features to create correlation plots for}
}
\description{
Creates a grid of correlation plots for a list of features and receptors.
Cells with dropout for the specific receptor are not included.
}
