% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fxns.R
\name{create_domino}
\alias{create_domino}
\title{Create a domino object and prepare it for network construction}
\usage{
create_domino(
  signaling_db,
  features,
  ser = NULL,
  z_scores = NULL,
  clusters = NULL,
  df = NULL,
  gene_conv = NULL,
  verbose = TRUE,
  use_complexes = TRUE
)
}
\arguments{
\item{signaling_db}{Path to directory of signaling database directory. The directory must include genes.csv, proteins.csv, interactions.csv, and complexes.csv formated according to cellphonedb2 syntax.}

\item{features}{Either a path to a csv containing cell level features of interest (ie. the auc matrix from pySCENIC) or named matrix with cells as columns and features as rows.}

\item{ser}{A Seurat object containing scaled RNA expression data in the RNA assay slot and cluster identity. Either a ser object OR z_scores and clusters must be provided. If ser is present z_scores and clusters will be ignored.}

\item{z_scores}{A matrix containing z-scored expression data for all cells with cells as columns and features as rows. Either z_scores and clusters must be provided OR a ser object. If ser is present z_scores and clusters will be ignored.}

\item{clusters}{A named factor containing cell cluster with names as cells. Either clusters and z_scores OR ser must be provided. If ser is present z_scores and clusters will be ignored.}

\item{df}{Optional. Either a path to discovered motifs from pySCENIC as a csv file or a data frame following the format of df.csv from pySCENIC}

\item{gene_conv}{Optional. Vector of length two containing some combination of 'ENSMUSG', 'ENSG', 'MGI', or 'HGNC' where the first vector is the current gene format in the database and the second is the gene format in the data set. If present, the function will use biomaRt to convert the database to the data sets gene format.}

\item{verbose}{Boolean indicating whether or not to print progress during computation.}

\item{use_complexes}{Boolean indicating whether you wish to use receptor/ligand complexes in the receptor ligand signaling database. This may lead to problems if genes which are preserved acrossed many functionally different signaling complexes are found highly expressed or correlated with features in your data set.}
}
\value{
A domino object.
}
\description{
This function reads in a receptor ligand signaling database, cell level 
features of some kind (ie. output from pySCENIC), z-scored single cell data, 
and cluster id for single cell data, calculates a correlation matrix between 
receptors and other features (this is transcription factor module scores if 
using pySCENIC), and finds features enriched by cluster. It will return a 
domino object prepared for build_domino, which will calculate a signaling 
network.
}
