% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{do_norm}
\alias{do_norm}
\title{Normalize a matrix to its max value by row or column}
\usage{
do_norm(mat, dir)
}
\arguments{
\item{mat}{Matrix to be normalized}

\item{dir}{Direction to normalize the matrix c('row', 'col')}
}
\value{
A normalized matrix in the direction specified.
}
\description{
Normalizes a matrix to its max value by row or column
}
\examples{
mat <- matrix(c(1, 3, 2, 5, 3, 5, 2, 1, 6), nrow = 3)
do_norm(mat, "row")

}
\keyword{internal}
