% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_correlations}
\alias{dom_correlations}
\title{Access correlations}
\usage{
dom_correlations(dom)
}
\arguments{
\item{dom}{A domino object that has been created with \code{\link[=create_domino]{create_domino()}}}
}
\value{
A matrix containing the correlation values for each receptor (row) by transcription factor (column)
}
\description{
A function to pull receptor-transcription factor correlations from a domino object
}
\examples{
cor_matrix <- dom_correlations(dominoSignal:::pbmc_dom_built_tiny)

}
