% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_de}
\alias{dom_de}
\title{Access differential expression}
\usage{
dom_de(dom)
}
\arguments{
\item{dom}{A domino object that has been created with \code{\link[=create_domino]{create_domino()}}}
}
\value{
A matrix containing the p values for differential expression of transcription factors (rows) in each cluster (columns)
}
\description{
A function to pull differential expression p values from a domino object
}
\examples{
de_mat <- dom_de(domino2:::pbmc_dom_built_tiny)

}
