% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_tf_activation}
\alias{dom_tf_activation}
\title{Access transcription factor activation}
\usage{
dom_tf_activation(dom)
}
\arguments{
\item{dom}{a domino object that has been created with \code{\link[=create_domino]{create_domino()}}}
}
\value{
A matrix containing the transcription factor activation scores for each TF (row) by cell (column)
}
\description{
A function to pull transcription factor activation scores from a domino object
}
\examples{
example(build_domino, echo = FALSE)
tf_activation <- dom_tf_activation(pbmc_dom_built_tiny)

}
