% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.diff.R
\name{domino.diff}
\alias{domino.diff}
\title{domino.diff}
\usage{
domino.diff()
}
\description{
Shows changes between your local version of project's data,
and the version uploaded to the Domino servers.
}
\keyword{diff}

