% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.snapshot.R
\name{domino.snapshot}
\alias{domino.snapshot}
\title{domino.snapshot}
\usage{
domino.snapshot(commitMessage)
}
\arguments{
\item{commitMessage}{A committ message to record with you snapshot}
}
\description{
take a snapshot of your command history and workspace.
}

