% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.status.R
\name{domino.status}
\alias{domino.status}
\title{domino.status}
\usage{
domino.status(...)
}
\arguments{
\item{...}{Arguments normally passed to the domino status command.}
}
\description{
Shows status of current Domino project.
}
\keyword{status}

