% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.sync.R
\name{domino.sync}
\alias{domino.sync}
\title{domino.sync}
\usage{
domino.sync()
}
\description{
Synchronizes the local project copy with the server project 
copy, by running 'download' followed by 'upload'.
}
\keyword{sync}

