\name{domino.upload}
\alias{domino.upload}
\encoding{UTF-8}
\title{
domino.upload
}
\description{
Uploads local version of Domino project files to the server.
}
\usage{
domino.upload(commitMessage)
}
\arguments{
  \item{commitMessage}{
Commit message for Domino client, explaining changes you did to the project code in recent update. Ex. \code{"updated project files"}
}
}
\keyword{ upload }


