% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fxns.R
\name{get_all_reclig}
\alias{get_all_reclig}
\title{Get All Ligands}
\usage{
get_all_reclig(dom, expressed_only = T, rec_lig = "rec")
}
\arguments{
\item{dom}{domino object}

\item{expressed_only}{logical indicating whether to subset ligands based on expression in dom@z_scores}
}
\value{
vector of ligands if expressed_only = T, list if F
}
\description{
Get all unique ligands present in dom@linkages$rec_lig
}
\details{
Get all unique ligands in a domino object, expanding all ligand complexes as well. Optionally subset by expression.
}
