% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fxns.R
\name{get_regulon_info}
\alias{get_regulon_info}
\title{Parse regulon information for the domino object}
\usage{
get_regulon_info(regulon_df)
}
\arguments{
\item{regulon_df}{Optional. Either a path to discovered motifs from pySCENIC as a csv file or a data frame following the format of df.csv from pySCENIC}
}
\value{
transcription factor target list object.
}
\description{
Parse regulon information for the domino object
}
