% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{ggplot_col_gen}
\alias{ggplot_col_gen}
\title{Generate ggplot colors}
\usage{
ggplot_col_gen(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
A vector of colors according to ggplot color generation.
}
\description{
Accepts a number of colors to generate and generates a ggplot color spectrum.
}
