% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{invert_rec_lig_expr}
\alias{invert_rec_lig_expr}
\title{Invert Receptor Ligand Data}
\usage{
invert_rec_lig_expr(dom)
}
\arguments{
\item{dom}{domino object built by domino build}
}
\value{
domino object with updated \code{rec_signaling} slot and \code{cr_signaling_matrices} slot
}
\description{
Reformat exisiting domino object to also record expression data/receptor mapping
from the ligand-receptor direction rather than receptor-ligand
}
