% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{invert_rec_lig_linkages}
\alias{invert_rec_lig_linkages}
\title{Invert Receptor Ligand Linkages}
\usage{
invert_rec_lig_linkages(dom)
}
\arguments{
\item{dom}{domino object built by domino build}
}
\value{
domino object with new entry in dom@linkages called \code{lig_rec}, which is a list
with one element per ligand and the list values are all receptors mapped to that ligand by rl_map
}
\description{
Use dom@misc$rl_map to create inverted version of dom@linkages$rec_lig
}
