% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fxns.R
\name{mean_ligand_expression}
\alias{mean_ligand_expression}
\title{Calculate mean ligand expression as a data.frame for plotting in circos plot}
\usage{
mean_ligand_expression(x, ligands, cell_ident, cell_barcodes, destination)
}
\arguments{
\item{x}{gene by cell expression matrix}

\item{ligands}{character vector of ligand genes to be quantified}

\item{cell_ident}{vector of cell type (identity) names for which to calculate mean ligand gene expression}

\item{cell_barcodes}{vector of cell barcodes (colnames of x) belonging to cell_ident to calculate mean expression across}

\item{destination}{name of the receptor with which each ligand interacts}
}
\value{
data frame of ligand expression targeting the specified receptor
}
\description{
Creates a data frame of mean ligand expression for use in plotting a circos
plot of ligand exression and saving tables of mean expression.
}
