% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mock_linkage_summary}
\alias{mock_linkage_summary}
\title{Create a mock linkage summary object}
\usage{
mock_linkage_summary()
}
\value{
obj a linkage summary object
}
\description{
Create a mock linkage summary object
}
