% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mockdata}
\alias{mockdata}
\title{dominoSignal object mocks}
\format{
A list of:
\describe{
\item{linkage_sum_tiny}{Mock of linkage summary}
\item{tiny_differential_linkage_c1}{An examlpe of diff linkage}
\item{tiny_differential_linkage_c2}{Another example of diff linkage}
}
}
\usage{
mockdata
}
\description{
A list of four mocks of dominoSignal data.
}
\keyword{datasets}
