% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_fxns.R
\name{outgoing_network}
\alias{outgoing_network}
\title{Create outgoing signaling network}
\usage{
outgoing_network(
  dom,
  outgoing_cluster = NULL,
  rec_clusters = NULL,
  plot_ligands = NULL,
  verbose = T
)
}
\arguments{
\item{dom}{domino object}

\item{outgoing_cluster}{vector of cluster names}

\item{plot_ligands}{vector of ligands to include}

\item{rec_cluster}{vector of cluster names}
}
\value{
table
}
\description{
Create outgoing signaling network
}
\details{
Make a table
}
