% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{PBMC3K single cell RNA sequencing data}
\format{
\subsection{\code{pbmc}}{

A Seurat object that has been through preliminary analysis, including normalization, scaling, PCA, UMAP, clustering, and cluster annotation
}
}
\source{
\url{https://cf.10xgenomics.com/samples/cell/pbmc3k/pbmc3k_filtered_gene_bc_matrices.tar.gz}
}
\usage{
pbmc
}
\description{
A small example single cell data set from 10x genomcs
}
\keyword{internal}
