% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_dom}
\alias{pbmc_dom}
\title{Domino object created from PBMC3K single cell RNA sequencing data}
\format{
\subsection{\code{pbmc_dom}}{

A domino object that has been built using SCENIC, CellphoneDB v4, and the domino2 package
}
}
\usage{
pbmc_dom
}
\description{
A small example domino object created with the \code{pbmc} data
}
\keyword{internal}
