% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_if_char}
\alias{read_if_char}
\title{Read in data if an object looks like path to it}
\usage{
read_if_char(obj)
}
\arguments{
\item{obj}{object to read if not already object}
}
\value{
Object itself or data read in from path
}
\description{
Read in data if an object looks like path to it
}
\keyword{internal}
