% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_complexes}
\alias{resolve_complexes}
\title{Convert between complex names and gene names}
\usage{
resolve_complexes(dom, genes)
}
\arguments{
\item{dom}{A domino object}

\item{genes}{A vector of genes, some of which may be complexes}
}
\value{
A list where any complexes are mapped to a vector of
component genes. The list names are set to the input gene names.
}
\description{
Convert between complex names and gene names
}
\keyword{internal}
