% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCENIC}
\alias{SCENIC}
\title{SCENIC AUC subset}
\format{
A list of:
\describe{
\item{auc_tiny}{A subset of SCENIC AUCs}
\item{regulons_tiny}{A subset of SCENIC regulons}
}
}
\source{
\url{https://zenodo.org/records/10951634/files}
}
\usage{
data("SCENIC")
}
\description{
A subset of SCENIC AUCs as applied to PBMC data.
}
\keyword{datasets}
