% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{signaling_heatmap}
\alias{signaling_heatmap}
\title{Create a network heatmap}
\usage{
signaling_heatmap(
  dom,
  clusts = NULL,
  min_thresh = -Inf,
  max_thresh = Inf,
  scale = "none",
  normalize = "none",
  ...
)
}
\arguments{
\item{dom}{A domino object with network built (build_domino)}

\item{clusts}{A vector of clusters to be included. If NULL then all clusters are used.}

\item{min_thresh}{Minimum signaling threshold for plotting. Defaults to -Inf for no threshold.}

\item{max_thresh}{Maximum signaling threshold for plotting. Defaults to Inf for no threshold.}

\item{scale}{How to scale the values (after thresholding). Options are 'none', 'sqrt' for square root, or 'log' for log10.}

\item{normalize}{Options to normalize the matrix. Normalization is done after thresholding and scaling. Accepted inputs are 'none' for no normalization, 'rec_norm' to normalize to the maximum value with each receptor cluster, or 'lig_norm' to normalize to the maximum value within each ligand cluster}

\item{...}{Other parameters to pass to NMF::aheatmap}
}
\description{
Creates a heatmap of the signaling network. Alternatively, the network 
matrix can be accessed directly in the signaling slot of a domino object.
}
