% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fxns.R
\name{table_convert_genes}
\alias{table_convert_genes}
\title{Convert Genes Using Table}
\usage{
table_convert_genes(genes, from, to, conversion_table)
}
\arguments{
\item{genes}{The genes to convert.}

\item{from}{Gene symbol type of the input (ENSG, ENSMUSG, HGNC, MGI)}

\item{to}{Desired gene symbol type for the output (HGNC, MGI)}

\item{conversion_table}{A data.frame with column names corresponding to gene symbol types (mm.ens, hs.ens, mgi, hgnc)
and rows corresponding to the gene symbols themselves}
}
\value{
Data frame of genes with original and corresponding converted symbols
}
\description{
Takes a vector of gene inputs and returns converted gene table
}
\examples{
conversion_table <- read.csv(
 "http://www.informatics.jax.org/downloads/reports/HOM_MouseHumanSequence.rpt",
 sep = "\t")
mgi_genes <- c("Ptprc", "Cd3e", "Cd8a", "Cd4", "Foxp3")
table_convert_genes(mgi_genes, "MGI", "HGNC", conversion_table)

}
\keyword{internal}
