% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg}
\alias{check_arg}
\title{Check input arguments}
\usage{
check_arg(
  arg,
  allow_class = NULL,
  allow_len = NULL,
  allow_range = NULL,
  allow_values = NULL,
  need_vars = c(NULL),
  need_colnames = FALSE,
  need_rownames = FALSE,
  need_names = FALSE
)
}
\arguments{
\item{arg}{The argument to check}

\item{allow_class}{Vector of allowed classes}

\item{allow_len}{Vector of allowed lengths}

\item{allow_range}{range of minimum and maximum values i.e. c(1, 5)}

\item{allow_values}{Vector of allowed values}

\item{need_vars}{Vector of required variables}

\item{need_colnames}{Logical for whether colnames are required}

\item{need_rownames}{Logical for whether rownames are required}

\item{need_names}{Logical for whether names are required}
}
\value{
logical
}
\description{
Accepts an object and rules for checking, stops if rules not met.
}
\keyword{internal}
