% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{cor_scatter}
\alias{cor_scatter}
\title{Create a correlation plot between transcription factor activation score and receptor}
\usage{
cor_scatter(dom, tf, rec, remove_rec_dropout = FALSE, ...)
}
\arguments{
\item{dom}{Domino object with network built (\code{\link[=build_domino]{build_domino()}})}

\item{tf}{Target TF module for plotting with receptor}

\item{rec}{Target receptor for plotting with TF}

\item{remove_rec_dropout}{Whether to remove cells with zero expression for plot. This should match the same setting as in build_domino.}

\item{...}{Other parameters to pass to ggscatter.}
}
\value{
a ggplot object
}
\description{
Create a correlation plot between transcription factor activation score and receptor
}
