% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fxns.R
\name{create_rl_map_cellphonedb}
\alias{create_rl_map_cellphonedb}
\title{Create a receptor-ligand map from a cellphonedb signaling database}
\usage{
create_rl_map_cellphonedb(
  genes,
  proteins,
  interactions,
  complexes = NULL,
  database_name = "CellPhoneDB",
  gene_conv = NULL,
  gene_conv_host = "https://www.ensembl.org"
)
}
\arguments{
\item{genes}{Dataframe or file path to table of gene names in uniprot, hgnc_symbol, or ensembl format in cellphonedb database format}

\item{proteins}{Dataframe or file path to table of protein features in cellphonedb format}

\item{interactions}{Dataframe or file path to table of protein-protein interactions in cellphonedb format}

\item{complexes}{Optional: dataframe or file path to table of protein complexes in cellphonedb format}
}
\value{
A data frame where each row describes a possible receptor-ligand interaction
}
\description{
Generates a data frame of ligand-receptor interactions from a CellPhoneDB database annotating the genes encoding the interacting ligands and receptors to be queried in transcriptomic data.
}
