% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{do_norm}
\alias{do_norm}
\title{Normalize a matrix to its max value by row or column}
\usage{
do_norm(matrix, dir)
}
\arguments{
\item{matrix}{The matrix to be normalized}

\item{dir}{The direction to normalize the matrix c('row', 'col')}
}
\value{
Normalized matrix in the direction specified.
}
\description{
Normalizes a matrix to its max value by row or column
}
