% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_zscores}
\alias{dom_zscores}
\title{Access z-scores}
\usage{
dom_zscores(dom)
}
\arguments{
\item{dom}{A domino object that has been created with \code{\link[=create_domino]{create_domino()}}}
}
\value{
A matrix containing the z-scored gene expression values for each gene (row) by cell (column)
}
\description{
A function to pull z-scored expression from a domino object
}
\examples{
zscores <- dom_zscores(domino2:::pbmc_dom_built_tiny)

}
