% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.init.R
\name{domino.init}
\alias{domino.init}
\title{domino.init}
\usage{
# inits a project inside current directory, with given name.
# ex. if run in ~/my_project, with "my_name" as a param,
# it will create a Domino project called my-param inside ~/my_project directory.
domino.init("projectName")
}
\arguments{
\item{projectName}{Project name for the project that will be created in 
current working directory.}
}
\description{
Initializes new domino project in current directory.
It can also set arbitrary name to the project,
even if different from current directory name.
}
\examples{
\dontrun{
# in directory ./
domino.init("my-new-project")
# new project with name "my-new-project" is initialized inside current directory.
}

}
\keyword{init}

