% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.run.R
\name{domino.run}
\alias{domino.run}
\title{domino.run}
\usage{
domino.run(..., publishApiEndpoint=FALSE)
}
\arguments{
\item{...}{All the run arguments will be joined together using space character.
Ex. \code{domino.run("main.py", "-xvz", "my-file1.csv")}}

\item{publishApiEndpoint}{Whether or not to republish the project's API endpoint at the end of the run.}
}
\description{
Runs your project on Domino servers with given parameters.
}
\examples{
\dontrun{
my_data <- 4
domino.run("main.R","1","my-file1.csv", my_data)
#=> Runs "domino main.R 1 my-file1.csv 4"
}
}

