% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.upload.R
\name{domino.upload}
\alias{domino.upload}
\title{domino.upload}
\usage{
domino.upload(commitMessage)
}
\arguments{
\item{commitMessage}{Commit message for Domino client, explaining changes you
did to the project code in recent update. Ex. \code{"updated project files"}}
}
\description{
Uploads local version of Domino project files to the server.
}
\keyword{upload}

