% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_dom}
\alias{pbmc_dom}
\title{Example domino object from pbmc3k data}
\format{
\subsection{\code{pbmc_dom}}{

A domino object based on the 10x Genomics pbmc3k data set that serves as an example in tutorials
and can also be used to explore the object structure and functionality.
}
}
\usage{
pbmc_dom
}
\description{
A small example single cell data set from 10x Genomics that has been through preliminary analysis,
including normalization, scaling, PCA, UMAP, clustering, and cluster annotation and then used in an example
domino2 analysis.
}
\keyword{datasets}
