% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{print,domino-method}
\alias{print,domino-method}
\title{Print domino object}
\usage{
\S4method{print}{domino}(x, ...)
}
\arguments{
\item{x}{Domino object}
}
\description{
Prints a summary of a domino object
}
\keyword{internal}
