% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_if_char}
\alias{read_if_char}
\title{Read in data if an object looks like path to it.}
\usage{
read_if_char(obj)
}
\arguments{
\item{obj}{Object to read if not already object}
}
\value{
obj Object itself in case its not a character
}
\description{
Read in data if an object looks like path to it.
}
\keyword{internal}
