% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fxns.R
\name{rename_clusters}
\alias{rename_clusters}
\title{Renames clusters in a domino object}
\usage{
rename_clusters(dom, clust_conv)
}
\arguments{
\item{dom}{Domino object to rename clusters in}

\item{clust_conv}{Named vector of conversions from old to new clusters. Values are taken as new clusters IDs and names as old cluster IDs.}
}
\value{
A domino object with clusters renamed in all applicable slots.
}
\description{
This function reads in a receptor ligand signaling database, cell level
features of some kind (ie. output from pySCENIC), z-scored single cell data,
and cluster id for single cell data, calculates a correlation matrix between
receptors and other features (this is transcription factor module scores if
using pySCENIC), and finds features enriched by cluster. It will return a
domino object prepared for \code{\link[=build_domino]{build_domino()}}, which will calculate a signaling
network.
}
\examples{
new_clust <- c("CD8_T_cell" = "CD8+ T Cells",
 "CD14_monocyte" = "CD14+ Monocytes", "B_cell" = "B Cells")
pbmc_dom_built_tiny <- rename_clusters(domino2:::pbmc_dom_built_tiny, new_clust)

}
\keyword{internal}
