% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fxns.R
\name{resolve_complexes}
\alias{resolve_complexes}
\title{Resolve Complexes}
\usage{
resolve_complexes(dom, genes)
}
\arguments{
\item{dom}{domino object}

\item{genes}{vector of complex names to resolve}
}
\value{
list of length(genes), with names(list) == genes. List values are the same as the names if not in a complex and are the complex genes if they are in a complex.
}
\description{
Expand any complex names into their component gene names
}
\details{
Ligand names (which are stored in dom@linkages$rec_lig) can refer to complexes that are
detailed in dom@linkages$complexes. Search all provided names and if the name is a complex, output component genes, otherwise output the name
}
