% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fxns.R
\name{resolve_names}
\alias{resolve_names}
\title{Resolve Names}
\usage{
resolve_names(dom, genes, rec_lig = "lig")
}
\arguments{
\item{dom}{domino object}

\item{genes}{vector of gene names to resolve}
}
\value{
vector of length(genes) with applicable values replaced using dom@misc$rl_map information
}
\description{
Return gene names of ligands with non-standard names
}
\details{
Ligand names (which are stored in dom@linkages$rec_lig) do not always match the gene name
Search all provided names and output the gene name if the ligand name is non-standard
}
