% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{show,domino-method}
\alias{show,domino-method}
\title{Show domino object information}
\usage{
\S4method{show}{domino}(object)
}
\arguments{
\item{object}{Domino object}
}
\value{
a printed description of the number of cells in a domino object and its build status
}
\description{
Shows content overview of domino object
}
\examples{
dominoSignal:::pbmc_dom_built_tiny

show(dominoSignal:::pbmc_dom_built_tiny)

}
\keyword{internal}
