% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_fxns.R
\name{summarize_linkages}
\alias{summarize_linkages}
\title{Summarize linkages from multiple domino objects}
\usage{
summarize_linkages(domino_results, subject_meta, subject_names = NULL)
}
\arguments{
\item{domino_results}{list of domino result with one domino object per subject. Names from the list must match subject_names.}

\item{subject_meta}{dataframe that includes the subject features by which the objects could be grouped. The first column should must be subject names}

\item{subject_names}{vector of subject names in domino_results. If NULL, defaults to first column of subject_meta.}
}
\value{
A linkage summary class object consisting of nested lists of the active transcription factors, active receptors, and incoming ligands for each cluster across multiple domino results.
}
\description{
Summarize linkages from multiple domino objects
}
